# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateCompanyDTO:
    properties:
      logo_url:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      plan_tier:
        enum:
        - free
        - premium
        - enterprise
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
      timezone:
        type: string
      trial_ends_at:
        type: string
    required:
    - name
    - plan_tier
    - slug
    - timezone
    type: object
...
  dtos.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dtos.RefreshTokenDTO:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  dtos.RegisterCompanyDTO:
    properties:
      admin_email:
        type: string
      admin_first_name:
        type: string
      admin_last_name:
        type: string
      admin_password:
        minLength: 8
        type: string
      company_name:
        type: string
      company_slug:
        type: string
      timezone:
        type: string
    required:
    - admin_email
    - admin_first_name
    - admin_last_name
    - admin_password
    - company_name
    - company_slug
    type: object
  dtos.RegisterDTO:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@dvra.com
    name: API Support
  description: API para sistema de reclutamiento y gestión de candidatos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DVRA API
  version: 1.2.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario y retorna tokens JWT
      parameters:
      - description: Credenciales de acceso
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Iniciar sesión
      tags:
      - Authentication
  /auth/me:
    get:
      consumes:
      - application/json
      description: Retorna la información del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener usuario actual
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Genera un nuevo access token usando el refresh token
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/dtos.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refrescar token de acceso
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registra un nuevo usuario. Usar RegisterCompany en su lugar
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Registrar nuevo usuario (deprecated)
      tags:
      - Authentication
  /auth/register-company:
    post:
      consumes:
      - application/json
      description: Crea una nueva empresa y su usuario administrador
      parameters:
      - description: Datos de empresa y admin
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterCompanyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Registrar empresa con admin
      tags:
      - Authentication
  /companies:
    get:
      consumes:
      - application/json
      description: Obtiene empresas (todas si es SuperAdmin, solo la actual si es
        usuario normal)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar empresas
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Crea una nueva empresa
      parameters:
      - description: Datos de la empresa
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCompanyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear empresa
      tags:
      - Companies
  /companies/{id}:
    get:
      consumes:
      - application/json
      description: Retorna información de una empresa específica
      parameters:
      - description: ID de la empresa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener empresa por ID
      tags:
      - Companies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"